rename/refactor viz to plotting
