+++ examples/plotting/README.rst
@@ -1,0 +1,8 @@
+.. _examples-plotting:
+
+Plotting
+========
+
+The :py:mod:`pulse2percept.plotting` module provides figures and animations
+that combine multiple pulse2percept objects. Individual objects retain their
+own ``plot()`` and ``play()`` methods.
