--- examples/viz/README.rst
@@ -1,9 +1,0 @@
-.. _examples-viz:
-
-Visualization
-=============
-
-The :py:mod:`pulse2percept.viz` module provides views that span more than one
-object. Stimuli, percepts, implants, and models each know how to draw
-themselves; ``viz`` puts them side by side, on a shared clock where they have
-one.
