+++ pulse2percept/plotting/data/argus1.png
