+++ pulse2percept/plotting/data/argus2.png
